+++ lk_qtquick_scaffold/launcher.py
@@ -3,9 +3,10 @@
 @Module  : launcher.py
 @Created : 2020-08-30
 @Updated : 2020-12-03
-@Version : 0.2.13
+@Version : 0.2.15
 @Desc    :
 """
+from os import path
 from PySide2.QtCore import QObject
 from PySide2.QtQml import QQmlApplicationEngine, QQmlContext
 from PySide2.QtWidgets import QApplication
@@ -21,48 +22,79 @@ class Application(QApplication):
             Python 垃圾回收机制误杀注入到 QML 全局变量的对象, 本身没有其他用处.
     """
     
-    def __init__(self, qmldir='', **kwargs):
+    def __init__(self, theme_dir='', **kwargs):
         """
         Args:
-            qmldir: 要引入外部自定义 QML 模块, 则填写该模块的父目录路径. 例如:
-                    |- myprj
+            theme_dir: 要引入外部自定义 QML 模块, 则填写该模块的父目录路径. 例如:
+                    |- my_prj
                         |- main.py
-                    |- scaffold
-                        |- ComponentLibX
-                            |- ComponentA.qml
-                            |- ComponentB.qml
-                            |- ComponentC.qml
+                    |- my_theme
+                        |- MyComponentLib
+                            |- MyComponentA.qml
+                            |- MyComponentB.qml
+                            |- MyComponentC.qml
                             |- qmldir  # The folder must include a 'qmldir' file
-                我们要在 myprj 中引入 scaffold 项目下的 ComponentLibX 组件库, 那
-                么:
-                    qmldir = '../scaffold'
+                我们要在 my_prj 中引入 my_theme 项目下的 MyComponentLib 组件库,
+                那么 theme_dir 参数填:
+                    theme_dir = '~/my_theme' (`~` 指你的绝对路径的目录)
                 
         Keyword Args:
-            organization (str): default 'dev.likianta.lk_qtquick_scaffold'
+            organization (str): default 'dev.likianta.lk_qtquick_scaffold'. 该键
+                是为了避免在 QML 中使用 QtQuick.Dialogs.FileDialog 时, 出现警告
+                信息:
+                    QML Settings: The following application identifiers have not
+                    been set: QVector("organizationName", "organizationDomain")
         """
         super().__init__()
-        
+
+        self.engine = QQmlApplicationEngine()
+        self.root = self.engine.rootContext()
+        self.__holder = {}
+
         # Set font to Microsoft Yahei if platform is Windows
         from platform import system
         if system() == 'Windows':
             from PySide2.QtGui import QFont
             self.setFont(QFont('Microsoft YaHei'))
+            
+        # Set organization name to avoid warning info if we use QtQuick.Dialogs.
+        # FileDialog component
         self.setOrganizationName(kwargs.get(
-            'organization', 'dev.likianta.lk_qtquick_scaffold'))
-        #   该步骤是为了避免在 QML 中使用 QtQuick.Dialogs.FileDialog 时, 出现警
-        #   告信息:
-        #       QML Settings: The following application identifiers have not
-        #       been set: QVector("organizationName", "organizationDomain")
+            'organization', 'dev.likianta.lk_qtquick_scaffold'
+        ))
         
-        self.engine = QQmlApplicationEngine()
-        self.root = self.engine.rootContext()
-        self.__holder = {}
+        # Register custom qml component library (folder's absolute paths)
+        # See `self.add_import_path` for detailed info.
+        self.curr_dir = path.dirname(__file__)  # -> '~/lk_qtquick_scaffold'
+        self.add_import_path(theme_dir or f'{self.curr_dir}/theme')
+        #   TODO: test whether '/theme/LightClean' worked as '/theme'
+        self.add_import_path(path.abspath(f'{self.curr_dir}/debugger'))
+        self.add_import_path(path.abspath(f'{self.curr_dir}/qml_helper'))
+        ''' Now you can use the following modules in Qml:
+            
+            import LKDebugger 1.0  // from '~/lk_qtquick_scaffold/debugger'
+            import LKHelper 1.0    // from '~/lk_qtquick_scaffold/qml_helper'
+            
+            import LightClean 1.0  // from '~/lk_qtquick_scaffold/theme' (if you
+                                   // didn't pass `theme_dir` your custom one,
+                                   // the built-in themes will be all available)
+            // PS: For now (2020-12) there is only one built-in theme provided.
+        '''
         
-        if not qmldir:
-            from os import path
-            qmldir = path.join(path.dirname(__file__), 'theme')
-            #   i.e. the abspath of './theme'.
-            #   TODO: test whether './theme/LightClean' worked as './theme'
+    def add_import_path(self, qmldir: str):
+        """
+        Args:
+            qmldir: The absolute path of directory, this dir should contain at
+                least one component library folder (the folder's first letter
+                shoule be capitalized), the the library folder should contain a
+                'qmldir' file (no suffix).
+                For example:
+                    qmldir = '~/lk_qtquick_scaffold/theme'
+                    # lk_qtquick_scaffold
+                    # |- theme (folder)  # <- we import this one's absolute path
+                    #    |- LightClean (folder)
+                    #       |- qmldir (file)
+        """
         self.engine.addImportPath(qmldir)
         
     def register_pyobj(self, obj: QObject, name=''):